apps/backend/auth_service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
firebase-admin==6.2.0
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.4.2